+++ checkpoint/orbax/checkpoint/__init__.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Defines exported symbols for the `orbax` packageorbax.checkpoint."""
+"""Defines exported symbols for Orbax Checkpointing."""
 
 import functools
 
@@ -43,3 +43,6 @@ from orbax.checkpoint.transform_utils import Transform
 # we can just use:
 #   `checkpointer = PyTreeCheckpointer()`
 PyTreeCheckpointer = functools.partial(Checkpointer, PyTreeCheckpointHandler())
+
+# A new PyPI release will be pushed everytime `__version__` is increased.
+__version__ = '0.1.1'