+++ export/orbax/export/__init__.py
@@ -12,9 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Orbax API."""
+"""Defines exported symbols for Orbax Export."""
 
-# Copybara: external sub-package imports.
+from orbax.export.dtensor_utils import dtensor_initialized
+from orbax.export.dtensor_utils import initialize_dtensor
+from orbax.export.dtensor_utils import maybe_enable_dtensor_export_on
+from orbax.export.export_manager import ExportManager
+from orbax.export.export_manager_base import ExportManagerBase
+from orbax.export.jax_module import JaxModule
+from orbax.export.serving_config import ServingConfig
 
 # A new PyPI release will be pushed everytime `__version__` is increased.
-__version__ = '0.1.1'
+__version__ = '0.0.1'